[Refactor] Color system 재구성
